requirements.txt
pandas==2.3.3
numpy==2.2.6
requests==2.32.5
faiss-cpu==1.12.0
matplotlib==3.10.7
streamlit-cropper==0.3.1
streamlit==1.51.0
torch==2.9.0
opencv-python
torchvision==0.24.0